Refactor common network plumbing into Octokit::Connection
This tries to share only what is truly common across Client,
EnterpriseAdminClient, and EnterpriseManagementConsoleClient and avoid
the heavy hammer that is class inheritance.
