+++ lib/octokit/client.rb
@@ -1,4 +1,6 @@
 require 'sawyer'
+require 'octokit/connection'
+require 'octokit/warnable'
 require 'octokit/arguments'
 require 'octokit/repo_arguments'
 require 'octokit/configurable'
@@ -53,6 +55,8 @@ module Octokit
 
     include Octokit::Authentication
     include Octokit::Configurable
+    include Octokit::Connection
+    include Octokit::Warnable
     include Octokit::Client::Authorizations
     include Octokit::Client::Commits
     include Octokit::Client::CommitComments
@@ -101,14 +105,6 @@ module Octokit
       login_from_netrc unless user_authenticated? || application_authenticated?
     end
 
-    # Compares client options to a Hash of requested options
-    #
-    # @param opts [Hash] Options to compare with current client options
-    # @return [Boolean]
-    def same_options?(opts)
-      opts.hash == options.hash
-    end
-
     # Text representation of the client, masking tokens and passwords
     #
     # @return [String]
@@ -127,126 +123,6 @@ module Octokit
       end
 
       inspected
-    end
-
-    # Make a HTTP GET request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Query and header params for request
-    # @return [Sawyer::Resource]
-    def get(url, options = {})
-      request :get, url, parse_query_and_convenience_headers(options)
-    end
-
-    # Make a HTTP POST request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Body and header params for request
-    # @return [Sawyer::Resource]
-    def post(url, options = {})
-      request :post, url, options
-    end
-
-    # Make a HTTP PUT request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Body and header params for request
-    # @return [Sawyer::Resource]
-    def put(url, options = {})
-      request :put, url, options
-    end
-
-    # Make a HTTP PATCH request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Body and header params for request
-    # @return [Sawyer::Resource]
-    def patch(url, options = {})
-      request :patch, url, options
-    end
-
-    # Make a HTTP DELETE request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Query and header params for request
-    # @return [Sawyer::Resource]
-    def delete(url, options = {})
-      request :delete, url, options
-    end
-
-    # Make a HTTP HEAD request
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Query and header params for request
-    # @return [Sawyer::Resource]
-    def head(url, options = {})
-      request :head, url, parse_query_and_convenience_headers(options)
-    end
-
-    # Make one or more HTTP GET requests, optionally fetching
-    # the next page of results from URL in Link response header based
-    # on value in {#auto_paginate}.
-    #
-    # @param url [String] The path, relative to {#api_endpoint}
-    # @param options [Hash] Query and header params for request
-    # @param block [Block] Block to perform the data concatination of the
-    #   multiple requests. The block is called with two parameters, the first
-    #   contains the contents of the requests so far and the second parameter
-    #   contains the latest response.
-    # @return [Sawyer::Resource]
-    def paginate(url, options = {}, &block)
-      opts = parse_query_and_convenience_headers(options.dup)
-      if @auto_paginate || @per_page
-        opts[:query][:per_page] ||=  @per_page || (@auto_paginate ? 100 : nil)
-      end
-
-      data = request(:get, url, opts)
-
-      if @auto_paginate
-        while @last_response.rels[:next] && rate_limit.remaining > 0
-          @last_response = @last_response.rels[:next].get
-          if block_given?
-            yield(data, @last_response)
-          else
-            data.concat(@last_response.data) if @last_response.data.is_a?(Array)
-          end
-        end
-
-      end
-
-      data
-    end
-
-    # Hypermedia agent for the GitHub API
-    #
-    # @return [Sawyer::Agent]
-    def agent
-      @agent ||= Sawyer::Agent.new(endpoint, sawyer_options) do |http|
-        http.headers[:accept] = default_media_type
-        http.headers[:content_type] = "application/json"
-        http.headers[:user_agent] = user_agent
-        if basic_authenticated?
-          http.basic_auth(@login, @password)
-        elsif token_authenticated?
-          http.authorization 'token', @access_token
-        elsif application_authenticated?
-          http.params = http.params.merge application_authentication
-        end
-      end
-    end
-
-    # Fetch the root resource for the API
-    #
-    # @return [Sawyer::Resource]
-    def root
-      get "/"
-    end
-
-    # Response for last HTTP request
-    #
-    # @return [Sawyer::Response]
-    def last_response
-      @last_response if defined? @last_response
     end
 
     # Duplicate client using client_id and client_secret as
@@ -312,94 +188,6 @@ module Octokit
     def client_secret=(value)
       reset_agent
       @client_secret = value
-    end
-
-    # Set Enterprise Management Console password
-    #
-    # @param value [String] Management console admin password
-    def management_console_password=(value)
-      reset_agent
-      @management_console_password = value
-    end
-
-    # Set Enterprise Management Console endpoint
-    #
-    # @param value [String] Management console endpoint
-    def management_console_endpoint=(value)
-      reset_agent
-      @management_console_endpoint = value
-    end
-
-    # Wrapper around Kernel#warn to print warnings unless
-    # OCTOKIT_SILENT is set to true.
-    #
-    # @deprecated Use Octokit.warn
-    # @return [nil]
-    def octokit_warn(*message)
-      Octokit.warn message
-    end
-
-    protected
-
-    def endpoint
-      api_endpoint
-    end
-
-    private
-
-    def reset_agent
-      @agent = nil
-    end
-
-    def request(method, path, data, options = {})
-      if data.is_a?(Hash)
-        options[:query]   = data.delete(:query) || {}
-        options[:headers] = data.delete(:headers) || {}
-        if accept = data.delete(:accept)
-          options[:headers][:accept] = accept
-        end
-      end
-
-      @last_response = response = agent.call(method, URI::Parser.new.escape(path.to_s), data, options)
-      response.data
-    end
-
-    # Executes the request, checking if it was successful
-    #
-    # @return [Boolean] True on success, false otherwise
-    def boolean_from_response(method, path, options = {})
-      request(method, path, options)
-      @last_response.status == 204
-    rescue Octokit::NotFound
-      false
-    end
-
-
-    def sawyer_options
-      opts = {
-        :links_parser => Sawyer::LinkParsers::Simple.new
-      }
-      conn_opts = @connection_options
-      conn_opts[:builder] = @middleware if @middleware
-      conn_opts[:proxy] = @proxy if @proxy
-      opts[:faraday] = Faraday.new(conn_opts)
-
-      opts
-    end
-
-    def parse_query_and_convenience_headers(options)
-      headers = options.fetch(:headers, {})
-      CONVENIENCE_HEADERS.each do |h|
-        if header = options.delete(h)
-          headers[h] = header
-        end
-      end
-      query = options.delete(:query)
-      opts = {:query => options}
-      opts[:query].merge!(query) if query && query.is_a?(Hash)
-      opts[:headers] = headers unless headers.empty?
-
-      opts
     end
   end
 end
