+++ lib/octokit/configurable.rb
@@ -96,6 +96,14 @@ module Octokit
     end
     alias setup reset!
 
+    # Compares client options to a Hash of requested options
+    #
+    # @param opts [Hash] Options to compare with current client options
+    # @return [Boolean]
+    def same_options?(opts)
+      opts.hash == options.hash
+    end
+
     def api_endpoint
       File.join(@api_endpoint, "")
     end
