+++ lib/octokit/enterprise_admin_client/license.rb
@@ -1,5 +1,5 @@
 module Octokit
-  class EnterpriseAdminClient < Octokit::Client
+  class EnterpriseAdminClient
 
     # Methods for the Enterprise License API
     #
