+++ lib/octokit/enterprise_admin_client/search_indexing.rb
@@ -1,5 +1,5 @@
 module Octokit
-  class EnterpriseAdminClient < Octokit::Client
+  class EnterpriseAdminClient
 
     # Methods for the Enterprise Search Indexing API
     #
